+++ cmd/root.go
@@ -30,12 +30,15 @@ import (
 )
 
 var (
-	version    = "0.0.11" // see CHANGELOG.md
+	version    = "0.0.13" // see CHANGELOG.md
 	verbose    bool       // global option
 	cfgFile    string     // global option
 	dryRun     bool       // global option
 	noCleanDst bool       // option for copy and mirror
 	skipHidden bool       // option for mirror and sync
+	// SFTP-specific
+	port             int
+	reverseDirection bool
 )
 
 // rootCmd represents the base command when called without any subcommands