+++ cmd/sync.go
@@ -41,18 +41,18 @@ import (
 )
 
 var syncCmd = &cobra.Command{
-	Use:     "sync",
+	Use:     "sync 'src' 'dst'",
 	Aliases: []string{"sy"},
 	Short:   "synchronize directory 'src' with directory 'dst'",
 	Long: `Synchronize the content of source directory with destination directory.
 Files will be copied if one is newer or doesn't exit in the destination.`,
-	Args: cobra.MaximumNArgs(2),
+	SilenceUsage: true,
+	Args:         cobra.MaximumNArgs(2),
 	RunE: func(_ *cobra.Command, args []string) error {
 		src := viper.GetString("src")
 		dst := viper.GetString("dst")
-
 		if len(args) < 2 && (src == "" || dst == "") {
-			return errors.New("missing required arguments [src] and [dst]")
+			return errors.New("missing required argument 'src' or 'dst'")
 		}
 
 		if len(args) == 2 {